# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Twitter service for Uala challenge",
        "title": "Uala Challenge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/tweet": {
            "post": {
                "description": "tweet to post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Tweet",
                "operationId": "createTweet",
                "parameters": [
                    {
                        "description": "Tweet Request",
                        "name": "tweet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createTweetDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.createTweetDTO": {
            "type": "object",
            "required": [
                "content",
                "user_id"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Hello!"
                },
                "user_id": {
                    "type": "string",
                    "example": "0f089136-3f38-4757-840c-d0c954782457"
                }
            }
        },
...
    }
}